ffmpeg resync. compile fix for firewirerecorder.cpp.
git-svn-id: http://svn.mythtv.org/svn/trunk@9483 7dbf422c-18fa-0310-86e9-fd20926502f2

+++ mythtv/libs/libmythtv/firewirerecorder.cpp
@@ -124,7 +124,7 @@ bool FirewireRecorder::Open(void)
      // Set buffered packets size
      size_t buffer_size = gContext->GetNumSetting("HDRingbufferSize",
                                                   50 * TSPacket::SIZE);
-     size_t buffered_packets = min(buffer_size / 4, kMaxBufferedPackets); 
+     size_t buffered_packets = min(buffer_size / 4, (size_t)kMaxBufferedPackets); 
      iec61883_mpeg2_set_buffers(fwmpeg, buffered_packets);
      VERBOSE(VB_IMPORTANT, LOC +
              QString("Buffered packets %1 (%2 KB)").
